+++ src/jalon04/client/client.c
@@ -16,82 +16,7 @@ jalon 03  : chat system
 #include <pthread.h>
 
 #include "../commons/functions.h"
-
-void get_addr_info(const char* address, const char* port, struct addrinfo** res) {
-
-	int status;
-	struct addrinfo hints, p;
-
-	memset(&hints,0,sizeof(hints));
-
-	hints.ai_family=AF_INET;
-	hints.ai_socktype=SOCK_STREAM;
-
-	hints.ai_flags = 0;
-	hints.ai_protocol = 0;
-
-
-  //  hints.ai_family =
-  //  hints.ai_socktype =
-
-	status = getaddrinfo(address,port,&hints,res);
-	if (status != 0) {
-		fprintf(stderr, "getaddrinfo: %s\n", gai_strerror(status));
-		exit(EXIT_FAILURE);
-	}
-
-}
-
-int do_socket(int domain, int type, int protocol) {
-	int sockfd;
-	int yes = 1;
-//create the socket
-	sockfd = socket(domain, type, protocol);
-
-//check for socket validity
-	if(sockfd==-1) {
-		fprintf(stderr,"Socket error\n");
-		return EXIT_FAILURE;
-	}
-
-// set socket option, to prevent "already in use" issue when rebooting the server right on
-
-	if (setsockopt(sockfd, SOL_SOCKET, SO_REUSEADDR, &yes, sizeof(int)) == -1)
-		error("ERROR setting socket options");
-
-	return sockfd;
-}
-
-void do_connect(int sockfd, const struct sockaddr *addr, socklen_t addrlen) {
-
-	int res = connect(sockfd, addr, addrlen);
-	if (res != 0) {
-		error("Connection error\n");
-	}
-
-	char buffer[1];
-	read(sockfd, buffer, 1);
-
-	if(strcmp(buffer,"0")==0) {
-		printf(" Server cannot accept incoming connections anymore. Try again later.\n");
-		close(sockfd);
-		exit(0);
-	}
-}
-
-void * readServ(void * arg){
-	int *sockfd = arg;
-	int maxlen = 500;
-	char * msg_s = malloc(sizeof(char) * maxlen);
-
-	for (;;){
-		memset(msg_s,0,maxlen);
-		//receive msg from server
-		readline(*sockfd, msg_s, maxlen);
-		printf("%s \n", msg_s);
-	}
-	pthread_exit(NULL);
-}
+#include "../commons/network.h"
 
 
 int main(int argc,char** argv)
